Initial version of burn request tests
